change output of action
Currently the action sets envvars as outputs to run tests.
This is easier for the customer, but we've seen more than a few times
users hitting Linux or GHA command size limits because of the amount
of tests to run.

To solve this problem, the present changes move the output to files
created by the action.
It requires a different - more complicated - command to run tests,
as indicated in `README.md`, but should still be doable.
And should solve the command size problem too.