chore: fail on error by default
Until know the '--fail-on-error' option for `create-commit` and
`create-report` commands was dependent on a `INPUTS_FAIL_ON_ERROR`.

However `INPUTS_FAIL_ON_ERROR` is not a valid input (see `action.yml`).
Plus, it's not smart to NOT activate this feature, because ATS needs
all these commands to succeed to actually work.

So the present changes activate '--fail-on-error' by default for
`create-commit` and `create-report` commands.